Start experimenting with a shared replayState across queries #33
