+++ src/read-models/members/get-failed-imports.ts
@@ -1,0 +1,12 @@
+/* eslint-disable unused-imports/no-unused-vars */
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import {DomainEvent} from '../../types';
+
+type FailedLinking = {
+  memberNumber: number;
+  email: string;
+};
+
+export const getFailedImports = (
+  events: ReadonlyArray<DomainEvent>
+): ReadonlyArray<FailedLinking> => [];
