+++ src/read-models/members/index.ts
@@ -1,7 +1,8 @@
 import {getAll, getAllDetails} from './get-all';
 import {getDetails} from './get';
 import {lookupByEmail} from './lookup-by-email';
-import {getFailedImports, getPotentialOwners} from './get-potential-owners';
+import {getPotentialOwners} from './get-potential-owners';
+import {getFailedImports} from './get-failed-imports';
 
 export const members = {
   lookupByEmail,
