+++ src/read-models/members/shared-state.ts
@@ -1,0 +1,81 @@
+import * as RA from 'fp-ts/ReadonlyArray';
+import {SubsetOfDomainEvent, filterByName} from '../../types';
+import {DomainEvent, EventName, isEventOfType} from '../../types/domain-event';
+import {pipe} from 'fp-ts/lib/function';
+import * as O from 'fp-ts/Option';
+import {Member} from './member';
+
+type Area = {
+  id: string;
+  owners: Set<number>;
+};
+
+type State = {
+  members: Map<Member['number'], Member>;
+  areas: Map<Area['id'], Area>;
+};
+
+const emptyState = (): State => ({
+  members: new Map(),
+  areas: new Map(),
+});
+
+const pertinentEventTypes: Array<EventName> = [
+  'MemberNumberLinkedToEmail',
+  'MemberDetailsUpdated',
+  'AreaCreated',
+  'OwnerAdded',
+  'OwnerAgreementSigned',
+];
+
+const handleEvent = (
+  state: State,
+  event: SubsetOfDomainEvent<typeof pertinentEventTypes>
+) => {
+  if (isEventOfType('MemberNumberLinkedToEmail')(event)) {
+    state.members.set(event.memberNumber, {
+      number: event.memberNumber,
+      email: event.email,
+      name: O.none,
+      pronouns: O.none,
+      agreementSigned: O.none,
+    });
+  }
+  if (isEventOfType('MemberDetailsUpdated')(event)) {
+    const current = state.members.get(event.memberNumber);
+    const name = O.fromNullable(event.name);
+    const pronouns = O.fromNullable(event.pronouns);
+    if (current) {
+      state.members.set(event.memberNumber, {...current, name, pronouns});
+    }
+  }
+  if (isEventOfType('OwnerAgreementSigned')(event)) {
+    const current = state.members.get(event.memberNumber);
+    if (current) {
+      state.members.set(event.memberNumber, {
+        ...current,
+        agreementSigned: O.some(event.signedAt),
+      });
+    }
+  }
+  if (isEventOfType('AreaCreated')(event)) {
+    state.areas.set(event.id, {id: event.id, owners: new Set()});
+  }
+  if (isEventOfType('OwnerAdded')(event)) {
+    const current = state.areas.get(event.areaId);
+    if (current) {
+      state.areas.set(event.areaId, {
+        ...current,
+        owners: current.owners.add(event.memberNumber),
+      });
+    }
+  }
+  return state;
+};
+
+export const replayState = (events: ReadonlyArray<DomainEvent>) =>
+  pipe(
+    events,
+    filterByName(pertinentEventTypes),
+    RA.reduce(emptyState(), handleEvent)
+  );
